DB lookup; (#266)
* DB lookup;

- add usecase for hash lookup, mode check and loading data from db;

* - fix tests;

+++ app/src/test/java/dgca/verifier/app/android/data/local/dcc/revocation/DccRevocationLocalDataSourceTest.kt
@@ -52,9 +52,9 @@ class DccRevocationLocalDataSourceTest {
 
     private val testDccRevocationPartition = DccRevocationPartition(
         kid = "a1b2c3",
-        x = 'a'.toByte(),
-        y = '1'.toByte(),
-        z = '2'.toByte(),
+        x = 'a',
+        y = '1',
+        z = '2',
         id = "id",
         expires = System.currentTimeMillis(),
         chunks = ""
